Implementation of move API calls, and attempt at serialization.
API calls in server proxy need to set url, set headers, set body, and post. I wrote helper functions for the headers and posting. They also need to throw an InvalidActionException. I added a message to that exception class, so we can tell the user what happened.

Serialization: here are my thoughts. Each model class is in charge of serializing itself and other associated things in it that don't have a class to serialize it (eg the TradeOffer would be in charge of serializing itself, and the acceptTrade JSON as well). I created a Message class that has a playerId and a message, and has serialization methods as an example.